updated curriculum

+++ src/app/services/curriculum.service.ts
@@ -1,72 +1,46 @@
-import * as express from 'express';
 import { ApiError } from '../../middlewares/ApiError';
-import UploadService from './upload.service';
-import StorageService from './storage.service';
 import Curriculum from '../models/curriculum.model';
 
 export default class CurriculumService {
-  public static async show(candidateId): Promise<Curriculum> {
-    const curriculum = await Curriculum.findOne({ where: { candidateId } });
-
+  public static async show(candidateId: number): Promise<Curriculum> {
+    const curriculum = await Curriculum.findByPk(candidateId);
     if (!curriculum) {
       throw new ApiError('Curriculo nao encontrado.', 404);
     }
-
     return curriculum;
   }
 
-  public static async store(body: express.Request, candidateId: number) {
-    const curriculumExists = await Curriculum.findOne({
-      where: { candidateId },
-    });
-
-    if (curriculumExists) {
-      throw new ApiError('candidado ja possui curriculo cadastrado', 400);
-    }
-
-    const { file } = await UploadService.store(body, 'Curriculum', {
-      allowExtenstions: ['.docx', '.doc', '.pdf'],
-      maxSize: 1024 * 2048,
-    });
-
+  public static async store(body): Promise<Curriculum> {
     const curriculum = await Curriculum.create({
-      filename: file.filename,
-      extension: file.originalname.substr(file.originalname.lastIndexOf('.')),
-      path: `/Curriculum/${file.filename}`,
-      candidateId,
+      ...body,
     }).catch((error) => {
       throw new ApiError(error, 400);
     });
 
-    return curriculum.save();
+    return curriculum;
   }
 
-  public static async update(candidateId: number, body) {
-    const { file } = await UploadService.store(body, 'Curriculum', {
-      allowExtenstions: ['.docx', '.doc', '.pdf'],
-      maxSize: 1024 * 2048,
+  public static async update(body, candidateId: number) {
+    const curriculum = await Curriculum.findByPk(candidateId);
+
+    if (!curriculum) {
+      throw new ApiError('Curriculo não localizado para o ID informado', 404);
+    }
+
+    await curriculum.update({ ...body }).catch((error) => {
+      throw new ApiError(error, 400);
     });
 
-    const curriculum = await Curriculum.findOne({ where: { candidateId } });
-
-    await StorageService.remove(curriculum.path);
-
-    await curriculum
-      .update({
-        filename: file.filename,
-        extension: file.originalname.slice(-4),
-        path: `/Curriculum/${file.filename}`,
-      })
-      .catch((error) => {
-        throw new ApiError(error, 400);
-      });
-
-    return curriculum.save();
+    return curriculum;
   }
 
   public static async delete(candidateId: number): Promise<any> {
-    const curriculum = await Curriculum.findOne({ where: { candidateId } });
-    StorageService.remove(curriculum.path);
+    const curriculum = await Curriculum.findByPk(candidateId);
+
+    if (!curriculum) {
+      throw new ApiError('Curriculo não localizado para o ID informado', 404);
+    }
+
     return curriculum.destroy({ force: true });
   }
 }
